+++ pkg/stacker/build.go
@@ -638,7 +638,6 @@ func runInternalGoSubcommand(config types.StackerConfig, args []string) error {
 		"--roots-dir", config.RootFSDir,
 		"--stacker-dir", config.StackerDir,
 		"--storage-type", config.StorageType,
-		"--internal-userns",
 	}
 
 	if config.Debug {
